requirements.txt
pandas
streamlit
requests
streamlit-lottie
seaborn
matplotlib
python-dotenv